Fixes for review 2.
